backend/pyproject.toml
[project]
name = "marketplacetracker"
version = "0.1.0"
description = "desc"
authors = [
    {name = "Ethan Katz"}
]
license = {text = "Apache 2.0"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "oracledb (>=3.4.2,<4.0.0)",
    "numpy (>=2.4.2,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "apify-client (>=2.4.1,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "ruff (>=0.15.0,<0.16.0)",
    "black (>=26.1.0,<27.0.0)",
    "types-requests (>=2.32.4.20260107,<3.0.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# tools

[tool.poetry]
package-mode = false

[tool.black]
line-length = 79
target-version = ["py310"]

[tool.ruff]
line-length = 79
target-version = "py310"
lint.select = ["ALL"]
lint.ignore = ["E501"]  # Black handles line length
src = ["backend"]

[tool.ruff.isort]
known-first-party = ["backend"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
